% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspicious.R
\name{suspect_benford}
\alias{suspect_benford}
\title{Provides a visual test of Benford's Law}
\usage{
suspect_benford(x)
}
\arguments{
\item{x}{a vector of class numeric}
}
\value{
a ggplot image with frequencies of initial numerals plotted against the theoretical
values given by Benford's Law
}
\description{
Benford's Law fails
Benford's Law is a theory which states that small digits (1, 2, 3) appear at the beginning of numbers much more
frequently than large digits (7, 8, 9). In theory Benford's Law can be used to detect anomalies in accounting
practices or election results, though in practice it can easily be misapplied. If you suspect a dataset has
been created or modified to deceive, Benford's Law is an excellent first test, but you should always verify
your results with an expert before concluding your data have been manipulated
}
