% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspicious.R
\name{suspect_dates}
\alias{suspect_dates}
\title{Check for Common Date Problems}
\usage{
suspect_dates(x)
}
\arguments{
\item{x}{a vector of class date or datetime}
}
\value{
an evaluation of whether the vector contains any suspicious dates
}
\description{
Check for Common Date Problems
Checks to determine whether any dates in a list are one of two problems: (1) are they the commonly-used
excel dates given when a date = 0 (which are different for windows and mac) or (2) the Unix Epoch of Jan
1, 1970. The former is given as a class date, the later as a class date-time.
}
