% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspicious.R
\name{suspect_frame_dimensions}
\alias{suspect_frame_dimensions}
\title{Check Dataframe Dimensions for common cut offs}
\usage{
suspect_frame_dimensions(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
an evaluation of whether the dataframe may have been cut off prior to import
}
\description{
Check Dataframe Dimensions for common cut offs
Spreadsheet has 65536 rows
The maximum number of rows an old-fashioned Excel spreadsheet was allowed to have was 65,536.
If you receive a dataset with that number of rows you have almost certainly been given
truncated data. Go back and ask for the rest. Newer versions of Excel allowed for 1,048,576
rows, so it's less likely you'll be working with data that hits the limit.
Spreadsheet has 255 columns
Apple's Numbers app can only handle spreadsheets with 255 columns, and the app will
truncate files that have more columns without warning the user. If you receive a dataset
with exactly 255 columns, ask if the file was ever opened or converted with Numbers.
}
