% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspicious.R
\name{suspect_locations}
\alias{suspect_locations}
\title{Check for Unusual Longitude/Latitude}
\usage{
suspect_locations(long, lat)
}
\arguments{
\item{long}{a vector of longitude values}

\item{lat}{a vector of latitude values}
}
\value{
an evaluation of whether the two vectors contain 0N 0E
}
\description{
Check for Unusual Longitude/Latitude
Simply checks if any columns of longitude and latitude data are simultaneously 0 and 0,
potentially indicating incorrectly coded data.
}
