% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspicious.R
\name{suspect_numeric}
\alias{suspect_numeric}
\title{Check for Common Bad Numbers}
\usage{
suspect_numeric(x)
}
\arguments{
\item{x}{a vector of class numeric}
}
\value{
an evaluation of whether the vector contains any suspicious numbers
}
\description{
Check for Common Bad Numbers
Checks for numbers that are associated with missing data e.g. 99999 or are
associated with encoding limits such as 255 or 65535
}
