% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspicious.R
\name{suspect_telephone}
\alias{suspect_telephone}
\title{Check for Fictitious Telephone Numbers}
\usage{
suspect_telephone(x)
}
\arguments{
\item{x}{a vector of phone numbers that may be stored as text.}
}
\value{
an evaluation of whether the vector contains any suspicious numbers
}
\description{
Check for Fictitious Telephone Numbers
Checks for phone numbers containing the sequence 555 (in any location), which is commonly
used to denote a fictitious number. Also checks whether the common faux number 867-5309
is within the vector.
}
