% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspicious.R
\name{suspect_zips}
\alias{suspect_zips}
\title{Check for Suspicious Zip Codes}
\usage{
suspect_zips(x)
}
\arguments{
\item{x}{a vector of zip codes}
}
\value{
an evaluation of whether the two vectors contain any of the four identified problems.
}
\description{
Check for Suspicious Zip Codes
Checks to determine whether potentially NA values are given (e.g. 99999) for a zip code, or
if 12345 or 90210 are given for a zip code.
}
